refactor yaml guard value seed
